my first script on c

+++ 0x00-hello_world/6-size.c
@@ -5,11 +5,11 @@
  */
 int main(void)
 {
-	printf("The size of an int is %lu bytes\n", sizeof(int));
-	printf("The size of a float is %lu bytes\n", sizeof(float));
-	printf("The size of a long long int is %lu bytes\n", sizeof(long long int));
-	printf("The size of a char is %lu bytes\n", sizeof(char));
-	printf("The size of a long int is %lu bytes\n", sizeof(long int));
+	printf("The size of an int is %lu bytes\n", (unsigned long)sizeof(int));
+	printf("The size of a float is %lu bytes\n", (unsigned long)sizeof(float));
+	printf("The size of a long long int is %lu bytes\n", (unsigned long)sizeof(long long int));
+	printf("The size of a char is %lu bytes\n", (unsigned long)sizeof(char));
+	printf("The size of a long int is %lu bytes\n", (unsigned long)sizeof(long int));
 
 	return (0);
 }
